requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.3
django-ckeditor-5==0.2.15
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.0.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
